fix prototypes

+++ bld/commonui/h/mem.h
@@ -33,8 +33,8 @@
 #define _MEM_H_INCLUDED
 
 void    MemFree( void *ptr );
-void    *MemAlloc( unsigned size );
-void    *MemReAlloc( void *ptr, unsigned size );
+void    *MemAlloc( size_t size );
+void    *MemReAlloc( void *ptr, size_t size );
 void    MemStart( void );
 
 #endif /* _MEM_H_INCLUDED */
